Resume Scanner using Keywords/requirements.txt
pdfminer.six
nltk
